Update README.md
Greetings! I'm Amoh Eric, a Data Alchemist with a fervent dedication to the realms of data science and blogging. My journey is encapsulated in a passionate story that unfolds on my blog Amoh Eric | Data Alchemist. Here, I share insights, strategies, and educational content on creating diverse online businesses and effective marketing strategies to thrive in the digital landscape.